Test filtering a list of files
To make it possible to test that little filter I had to add an
interface (fine) but a bunch of code to copy slice of objects into a
slice of interfaces (less fine). To make the GitHubfile structure match
the NamedFile interface I had to make its Name field lowercase (private). Also
fine, but then that meant it didn't marshal from JSON, so I had to add
an internal struct with exported fields. Whew.