Move everything that's only used by backfillUrns
Anything that isn't shared is moved to backfill.go. This is kind of
misleading since all the code is in the same package anyway, but as part
of this I did need to move the PreRunPersistent in checkMigrationsCmd to
be PreRun so that both the logging and initGithubParams functions are called.