Automate checking migrations via Travis
Add validatemigrationorder to Makefile. The name checkmigrations has other,
similar names in the Makefile related to generating migrations, so don't
use that.

Add an executable target to the Makefile, and depend on that for
validatemigrationorder. This is more make appropriate.

Make the build task always rebuild the executable, to match existing behavior.