VDB-1756: Update urn snapshot trigger functions (#419)
* Update urn snapshot trigger

We're determing which snapshots to remove by looking at their block
number - previously we had tried to get the block number from the
header. However, since the header being removed kicks off the series of
triggers that this is a part of, the header record is not longer in the
database to get the block number from. So, now we're looking at the urn
snapshot's diff, which is not deleted and does have the block height.

* Update migration timestamp

* Add a test that mimics the issue from VDB-1756